+++ tests/test_api_transfer_value.py
@@ -3,7 +3,7 @@ import unittest
 
 from icxcli.icx.utils import get_tx_hash, sign
 from icxcli.icx import wallet, FilePathIsWrong, PasswordIsWrong, NoEnoughBalanceInWallet, TransferFeeIsInvalid, \
-    AddressIsWrong
+    AddressIsWrong, FeeIsBiggerThanAmount, AmountIsInvalid, AddressIsSame
 TEST_DIR = os.path.dirname(os.path.abspath(__file__))
 url = 'https://testwallet.icon.foundation/api/'
 
@@ -48,14 +48,14 @@ class TestAPI(unittest.TestCase):
         """
 
         # Given
-        password = "1234"
+        password = "ejfnvm1234*"
         file_path = os.path.join(TEST_DIR, "test_keystore_for_transfer.txt")
 
         # When
         try:
             ret = bool(wallet.transfer_value_with_the_fee(
-                password, 0.01, 18, to="hxa23651905dfa12221dd36b860dc114ef7f7a0786",
-                amount="1", file_path=file_path, url=url))
+                password, 10000000000000000, to="hxa974f512a510299b53c55535c105ed962fd01ee2",
+                amount="1000000000000000000", file_path=file_path, url=url))
 
             # Then
             self.assertEqual(True, ret)
@@ -68,17 +68,19 @@ class TestAPI(unittest.TestCase):
         Case when key_store_file_path is wrong.
         """
         # Given
-        password = "1234"
+        password = "ejfnvm1234*"
         file_path = os.path.join(TEST_DIR, "test_keystore_for_transfer.txt")
         # When
         try:
             ret = wallet.transfer_value_with_the_fee(
-                password, 0.01, 18, to="hxa23651905dfa12221dd36b860dc114ef7f7a0786",
-                amount="1", file_path='./wrong_path', url=url)
+                password, 10000000000000000, to="hxa974f512a510299b53c55535c105ed962fd01ee2",
+                amount="10000000000000000000", file_path='./wrong_path', url=url)
             # Then
 
         except FilePathIsWrong:
             self.assertTrue(True)
+        else:
+            self.assertTrue(False)
 
     def test_transfer_case2(self):
         """Test for transfer_value_with_the_fee function.
@@ -90,64 +92,164 @@ class TestAPI(unittest.TestCase):
         # When
         try:
             ret = wallet.transfer_value_with_the_fee(
-                password, 0.01, 18, to="hxa23651905dfa12221dd36b860dc114ef7f7a0786",
-                amount="1", file_path=file_path, url=url)
+                password, 10000000000000000, to="hxa974f512a510299b53c55535c105ed962fd01ee2",
+                amount="1000000000000000000", file_path=file_path, url=url)
             # Then
 
         except PasswordIsWrong:
             self.assertTrue(True)
+        else:
+            self.assertTrue(False)
 
     def test_transfer_case3(self):
         """Test for transfer_value_with_the_fee function.
         Case when wallet does not have enough balance.
         """
         # Given
-        password = "1234"
+        password = "ejfnvm1234*"
         file_path = os.path.join(TEST_DIR, "test_keystore_for_transfer.txt")
         # When
         try:
             ret = wallet.transfer_value_with_the_fee(
-                password, 0.01, 18, to="hxa23651905dfa12221dd36b860dc114ef7f7a0786",
-                amount="100000000", file_path=file_path, url=url)
+                password, 10000000000000000, to="hxa974f512a510299b53c55535c105ed962fd01ee2",
+                amount="10000000000000000000000000000000000000000000000000", file_path=file_path, url=url)
             # Then
 
         except NoEnoughBalanceInWallet:
             self.assertTrue(True)
+        else:
+            self.assertTrue(False)
 
     def test_transfer_case4(self):
         """Test for transfer_value_with_the_fee function.
         Case when transfer fee is invalid.
         """
         # Given
-        password = "1234"
+        password = "ejfnvm1234*"
         file_path = os.path.join(TEST_DIR, "test_keystore_for_transfer.txt")
         # When
         try:
             ret = wallet.transfer_value_with_the_fee(
-                password, -1, 18, to="hxa23651905dfa12221dd36b860dc114ef7f7a0786",
-                amount="1000", file_path=file_path, url=url)
+                password, 100000000000, to="hxa974f512a510299b53c55535c105ed962fd01ee2",
+                amount="1000000000000000000", file_path=file_path, url=url)
             # Then
 
         except TransferFeeIsInvalid:
             self.assertTrue(True)
+        else:
+            self.assertTrue(False)
 
     def test_transfer_case5(self):
         """Test for transfer_value_with_the_fee function.
         Case when wallet address is wrong.
         """
         # Given
-        password = "1234"
+        password = "ejfnvm1234*"
         file_path = os.path.join(TEST_DIR, "test_keystore_for_transfer.txt")
-        print()
+
         # When
         try:
             ret = wallet.transfer_value_with_the_fee(
-                password, 0.01, 18, to="hxa23651905d221dd36b",
-                amount="1", file_path=file_path, url=url)
+                password, 10000000000000000, to="hxa974f512a510299b53c55535c105ed9",
+                amount="1000000000000000000", file_path=file_path, url=url)
             # Then
 
         except AddressIsWrong:
             self.assertTrue(True)
+        else:
+            self.assertTrue(False)
+
+    def test_transfer_case6(self):
+        """Test for transfer_value_with_the_fee function.
+        Case when Fee is not 10000000000000000.
+        """
+        password = "ejfnvm1234*"
+        file_path = os.path.join(TEST_DIR, "test_keystore_for_transfer.txt")
+        print()
+
+        try:
+            ret = wallet.transfer_value_with_the_fee(
+                password, 100000000000000, to="hxa974f512a510299b53c55535c105ed962fd01ee2",
+                amount="11234440000000000000", file_path=file_path, url=url)
+
+        except TransferFeeIsInvalid:
+            self.assertTrue(True)
+        else:
+            self.assertTrue(False)
+
+    def test_transfer_case7(self):
+        """Test for transfer_value_with_the_fee function.
+        Case when Fee is bigger than Amount.
+        """
+
+        password = "ejfnvm1234*"
+        file_path = os.path.join(TEST_DIR, "test_keystore_for_transfer.txt")
+        try:
+            ret = wallet.transfer_value_with_the_fee(
+                password, 10000000000000000, to="hxa974f512a510299b53c55535c105ed962fd01ee2",
+                amount="1000000000000000", file_path=file_path, url=url)
+
+        except FeeIsBiggerThanAmount:
+            self.assertTrue(True)
+        else:
+            self.assertTrue(False)
+
+    def test_transfer_case8(self):
+        """Test for transfer_value_with_the_fee function.
+        Case when Amount is 0.
+        """
+        password = "ejfnvm1234*"
+        file_path = os.path.join(TEST_DIR, "test_keystore_for_transfer.txt")
+        try:
+            ret = wallet.transfer_value_with_the_fee(
+                password, 10000000000000000, to="hxa974f512a510299b53c55535c105ed962fd01ee2",
+                amount="0", file_path=file_path, url=url)
+
+        except AmountIsInvalid:
+            self.assertTrue(True)
+        else:
+            self.assertTrue(False)
+
+    def test_transfer_case9(self):
+        """Test for transfer_value_with_the_fee function.
+        Case when balance is same as sum of Amount and Fee.
+        """
+
+        password = "ejfnvm1234*"
+        file_path = os.path.join(TEST_DIR, "test_keystore_for_transfer.txt")
+
+        try:
+            ret = wallet.transfer_value_with_the_fee(
+                password, 10000000000000000, to="hx95e12b1f98f9b847175849f51bed5d121e742f6a",
+                amount="1020000000000000000", file_path=file_path, url=url)
+
+            password = "Adas21312**"
+            file_path = os.path.join(TEST_DIR, "test_keystore_for_transfer2.txt")
+            ret = wallet.transfer_value_with_the_fee(
+                password, 10000000000000000, to="hx66425784bfddb5b430136b38268c3ce1fb68e8c5",
+                amount="1000000000000000000", file_path=file_path, url=url)
+
+        except AmountIsInvalid:
+            self.assertTrue(False)
+        else:
+            self.assertTrue(True)
+
+    def test_transfer_case10(self):
+        """Test for transfer_value_with_the_fee function.
+        Case when wallet address to transfer is same as wallet address to be sent.
+        """
+
+        password = "ejfnvm1234*"
+        file_path = os.path.join(TEST_DIR, "test_keystore_for_transfer.txt")
+        try:
+            ret = wallet.transfer_value_with_the_fee(
+                password, 10000000000000000, to="hx66425784bfddb5b430136b38268c3ce1fb68e8c5",
+                amount="0", file_path=file_path, url=url)
+
+        except AddressIsSame:
+            self.assertTrue(True)
+        else:
+            self.assertTrue(False)
 
 
 if __name__ == "__main__":