+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 # Changelog
 
+## [0.0.2] - March 28, 2018
+### Added
 
-## [0.0.1] - 2018-03-13
+### Changed
+* Use ```loop```, 10<sup>18</sup> icx , for value in ```transfer``` command, not ICX. 
+* Removed ```-d``` in ```transfer``` command. 
+
+### Removed
+N/A
+
+
+
+## [0.0.1] - March 13, 2018
 ### Added
  - Initialize version.